finished loading methods

+++ src/interfaces/GateInterface.java
@@ -3,7 +3,7 @@ package interfaces;
 public interface GateInterface {
 	void open();
 	void close();
-	void setOpenTime(int seconds);
-	int getOpenTime();
+	void setOpenTime(double seconds);
+	double getOpenTime();
 	boolean isOpen();
 }
